+++ Cargo.toml
@@ -66,7 +66,6 @@ members = [
     "acpi_tables",
     "api_client",
     "arch",
-    "arch_gen",
     "block_util",
     "devices",
     "event_monitor",
